+++ dutil-basic/src/main/java/com/dwarfeng/dutil/basic/time/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * 有关于时间的工具类以及方法。
+ *
+ * @author DwArFeng
+ * @since 0.4.2.a-beta
+ */
+package com.dwarfeng.dutil.basic.time;